requirements.txt
streamlit
tensorflow==2.15.0
torch
torchvision
opencv-python-headless
numpy
pillow
gdown